requirements.txt
aiogram==2.16
python-dotenv==0.19.1
